[ACS-6630] Added models, APIs and docs to js-api for extension-manager feature
