+++ lib/extensions/src/lib/config/action.extensions.ts
@@ -47,9 +47,3 @@ export interface ContentActionRef extends ExtensionElement {
         visible?: string;
     };
 }
-
-export interface ActionRef {
-    id: string;
-    type: string;
-    payload?: any;
-}
