+++ lib/extensions/src/lib/services/extension.service.spec.ts
@@ -18,9 +18,7 @@
 import { ExtensionService } from './extension.service';
 import { ExtensionLoaderService } from './extension-loader.service';
 import { ExtensionConfig } from '../config/extension.config';
-import { RuleRef } from '../config/rule.extensions';
-import { RouteRef } from '../config/routing.extensions';
-import { ActionRef } from '../config/action.extensions';
+import { ActionRef, RouteRef, RuleRef } from '@alfresco/js-api';
 import { ComponentRegisterService } from './component-register.service';
 import { RuleService } from './rule.service';
