+++ lib/extensions/src/lib/services/extension.service.ts
@@ -16,17 +16,16 @@
  */
 
 import { Injectable, Type, InjectionToken, Inject } from '@angular/core';
-import { RuleEvaluator, RuleRef, RuleContext } from '../config/rule.extensions';
+import { RuleEvaluator, RuleContext } from '../config/rule.extensions';
 import { ExtensionConfig } from '../config/extension.config';
 import { ExtensionLoaderService } from './extension-loader.service';
-import { RouteRef } from '../config/routing.extensions';
-import { ActionRef } from '../config/action.extensions';
 import * as core from '../evaluators/core.evaluators';
 import { ComponentRegisterService } from './component-register.service';
 import { RuleService } from './rule.service';
 import { ExtensionElement } from '../config/extension-element';
 import { BehaviorSubject, Observable } from 'rxjs';
 import { mergeArrays, mergeObjects } from '../config/extension-utils';
+import { ActionRef, RouteRef, RuleRef } from '@alfresco/js-api';
 
 /**
  * The default extensions factory
