+++ lib/extensions/src/lib/services/rule.service.ts
@@ -16,9 +16,10 @@
  */
 
 import { Injectable } from '@angular/core';
-import { RuleRef, RuleContext, RuleEvaluator, RuleParameter } from '../config/rule.extensions';
+import { RuleContext, RuleEvaluator } from '../config/rule.extensions';
 import { ExtensionConfig } from '../config/extension.config';
 import { ExtensionLoaderService } from './extension-loader.service';
+import { RuleParameter, RuleRef } from '@alfresco/js-api';
 
 @Injectable({
     providedIn: 'root'
