+++ lib/js-api/src/api/content-rest-api/model/extensionConfig.ts
@@ -1,6 +1,6 @@
 /*!
  * @license
- * Copyright © 2005-2025 Hyland Software, Inc. and its affiliates. All rights reserved.
+ * Copyright © 2005-2023 Hyland Software, Inc. and its affiliates. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,29 +15,17 @@
  * limitations under the License.
  */
 
-import { ActionRef, RouteRef, RuleRef } from '@alfresco/js-api';
+import { AppConfigPluginRef } from './appConfigPluginRef';
+import { RuleRef } from './ruleRef';
+import { RouteRef } from './routeRef';
+import { ActionRef } from './actionRef';
 
-export interface ExtensionRef {
-    $id: string;
-    $name: string;
-    $version: string;
-    $vendor: string;
-    $license: string;
-    $runtime?: string;
-    $description?: string;
-    $dependencies?: Array<string>;
-
-    appConfig?: any;
+export class ExtensionConfig {
+    appConfig: AppConfigPluginRef;
     rules?: Array<RuleRef>;
     routes?: Array<RouteRef>;
     actions?: Array<ActionRef>;
-
     features?: {
         [key: string]: any;
     };
-}
-
-export interface ExtensionConfig extends ExtensionRef {
-    $references?: Array<string | ExtensionRef>;
-    $ignoreReferenceList?: Array<string>;
 }
