+++ lib/js-api/src/api/content-rest-api/model/index.ts
@@ -23,10 +23,12 @@ export * from './actionDefinitionListList';
 export * from './actionExecResult';
 export * from './actionExecResultEntry';
 export * from './actionParameterDefinition';
+export * from './actionRef';
 export * from './activity';
 export * from './activityEntry';
 export * from './activityPaging';
 export * from './activityPagingList';
+export * from './appConfigPluginRef';
 export * from './agent';
 export * from './agentEntry';
 export * from './agentPaging';
@@ -79,6 +81,8 @@ export * from './download-status';
 export * from './downloadBodyCreate';
 export * from './downloadEntry';
 export * from './errorError';
+export * from './extensionConfig';
+export * from './extensionConfigEntry';
 export * from './favorite';
 export * from './favoriteBodyCreate';
 export * from './favoriteEntry';
@@ -156,6 +160,8 @@ export * from './renditionEntry';
 export * from './renditionPaging';
 export * from './renditionPagingList';
 export * from './revertBody';
+export * from './routeRef';
+export * from './ruleRef';
 export * from './sharedLink';
 export * from './sharedLinkBodyCreate';
 export * from './sharedLinkBodyEmail';
